Removing the update-dist feature Restarting ubuntu1604 after installing vmware tools

+++ script/vmware.sh
@@ -21,14 +21,16 @@ if [[ $PACKER_BUILDER_TYPE =~ vmware ]]; then
         rm -rf /tmp/vmware-tools-patches
     fi
 
-    # Recompile when the kernel is updated
-    # Make sure the kernel module is loaded at boot
-    echo "answer AUTO_KMODS_ENABLED yes" >> /etc/vmware-tools/locations
-    echo "vmhgfs" > /etc/modules-load.d/vmware.conf
-    
     VMWARE_TOOLBOX_CMD_VERSION=$(vmware-toolbox-cmd -v)
     echo "==> Installed VMware Tools ${VMWARE_TOOLBOX_CMD_VERSION}"
-
-    service ssh stop
-    reboot
+    
+    if [[ "$VMWARE_RESTART_AFTER_INSTALL" == true ]]; then
+        echo "==> Restarting to enable VMware tools"
+        # Recompile when the kernel is updated
+        # Make sure the kernel module is loaded at boot
+        echo "answer AUTO_KMODS_ENABLED yes" >> /etc/vmware-tools/locations
+        echo "vmhgfs" > /etc/modules-load.d/vmware.conf
+        service ssh stop
+        reboot
+    fi
 fi
